Verificações
Adicionando verificações sobre os dados enviados para o usuário dentro de index, além de também indicar os tipos dos parâmetros dos métodos. foram refeitas alguns códigos em TaskDAO.php para uma melhor expressão de erros pela API.

+++ API/DAO/TaskDAO.php
@@ -19,7 +19,6 @@ use Model\Task;
             $stmt->bindValue(2, $model->getTitulo());
             $stmt->bindValue(3, $model->getDescricao());
             $stmt->execute();
-
             $model->setId(parent::$conexao->lastInsertId());
             
             return $model;
@@ -32,7 +31,7 @@ use Model\Task;
             $stmt->bindValue(1, $id);
             $stmt->execute();
 
-            return $stmt->fetchObject("Model\Task");
+            return $stmt->fetchObject("Model\Task") ?: null;
         }
 
         public function selectAll(): array {
@@ -44,24 +43,40 @@ use Model\Task;
             return $stmt->fetchAll(DAO::FETCH_CLASS, "Model\Task");
         }
 
-        public function update(Task $model) : Task{
-             $sql = "UPDATE lembretes SET titulo = ?, descricao = ? WHERE id = ?";
+        public function update(Task $model) : ?bool{
+            $sql = "SELECT id from lembretes WHERE id = ?";
+            $stmt = parent::$conexao->prepare($sql);
+            $stmt->bindValue(1, $model->getId());
+            $stmt->execute();
+            $existe = $stmt->fetchColumn();
 
+            if(!$existe){
+                return false;
+            }
+            
+            $sql = "UPDATE lembretes SET titulo = ?, descricao = ? WHERE id = ?";
             $stmt = parent::$conexao->prepare($sql);
             $stmt->bindValue(1, $model->getTitulo());
             $stmt->bindValue(2, $model->getDescricao());
             $stmt->bindValue(3, $model->getId());
-            $stmt->execute();
-
-            $model->setId(parent::$conexao->lastInsertId());
+            $stmt->execute();       
             
-            return $model;
+            return true;
         }
 
         
         public function delete(int $id) : bool {
-            $sql = "DELETE FROM lembretes WHERE id = ?";
+            $sql = "SELECT id from lembretes WHERE id = ?";
+            $stmt = parent::$conexao->prepare($sql);
+            $stmt->bindValue(1, $id);
+            $stmt->execute();
+            $existe = $stmt->fetchColumn();
 
+            if(!$existe){
+                return false;
+            }
+
+            $sql = "DELETE FROM lembretes WHERE id = ?";
             $stmt = parent::$conexao->prepare($sql);
             $stmt->bindValue(1, $id);
             return $stmt->execute();
